refactor(ors_control_app): split routes/regions/management.ts into catalog/registry/provisioning
